+++ host/src/FunctionsNetHost/Logger.cs
@@ -22,7 +22,7 @@ namespace FunctionsNetHost
         {
             get
             {
-                return string.Equals(EnvironmentUtils.GetValue(EnvironmentSettingNames.FunctionsNetHostTrace), "1");
+                return string.Equals(EnvironmentUtils.GetValue(EnvironmentVariables.FunctionsNetHostTrace), "1");
             }
         }